move email utils to root directory
